+++ src/class-jetpack-importer.php
@@ -13,6 +13,15 @@ class Jetpack_Importer {
     {
         ?>
         <div class="wrap" style="max-width: 820px;">
+
+            <?php if (isset($_GET['success']) && $_GET['success'] == 0) { ?>
+                <div class="notice notice-error is-dismissible"><p>Sorry, something went wrong in trying to retrieve your stats from WordPress.com. Please check whether your API key and blog URL are correct.</p></div>
+            <?php } ?>
+
+            <?php if (isset($_GET['success']) && $_GET['success'] == 1) { ?>
+                <div class="notice notice-success is-dismissible"><p>Big success! Your stats are now imported into Koko Analytics.</p></div>
+            <?php } ?>
+
             <h1>Import analytics from JetPack Stats</h1>
             <p>To import your historical analytics data from JetPack Stats into Koko Analytics, provide your WordPress.com API key in the field below.</p>
 
@@ -65,20 +74,39 @@ class Jetpack_Importer {
     public function start_import(): void {
         // save params
         $params = [
-            'wpcom-api-key' => trim($_POST['wpcom-api-key']),
-            'wpcom-blog-uri' => trim($_POST['wpcom-blog-uri']),
-            'date-start' => trim($_POST['date-start']),
-            'date-end' =>trim($_POST['date-end']),
+            'wpcom-api-key' => trim($_POST['wpcom-api-key'] ?? ''),
+            'wpcom-blog-uri' => trim($_POST['wpcom-blog-uri'] ?? ''),
+            'date-start' => trim($_POST['date-start'] ?? ''),
+            'date-end' =>trim($_POST['date-end'] ?? ''),
         ];
-        update_option('koko_analytics_jetpack_import_params', $params, false);
+
+        // all params are required
+        if ($params['wpcom-api-key'] === '' || $params['wpcom-blog-uri'] === '' || $params['date-start'] === '' || $params['date-end'] === '') {
+            wp_safe_redirect(get_admin_url(null, '/index.php?page=koko-analytics&tab=jetpack_importer&success=0'));
+            exit;
+        }
 
         // first chunk is 30 days after date-start
-        $date_start = new \DateTimeImmutable($params['date-start']);
-        $date_end = new \DateTimeImmutable($params['date-end']);
+        try {
+            $date_start = new \DateTimeImmutable($params['date-start']);
+            $date_end = new \DateTimeImmutable($params['date-end']);
+            if ($date_end < $date_start) {
+                throw new \Exception("End date must be after start date");
+            }
+        } catch (\Exception $e) {
+            wp_safe_redirect(get_admin_url(null, '/index.php?page=koko-analytics&tab=jetpack_importer&success=0'));
+            exit;
+        }
+
+        // params are valid; let's go!
+        update_option('koko_analytics_jetpack_import_params', $params, false);
+
+        // work backwards from end date, so most recent stats first
+        $chunk_end = $date_end;
         $chunk_size = 30;
-        $chunk_end = $date_start->modify("+{$chunk_size} days");
-        if ($chunk_end > $date_end) {
-            $chunk_end = $date_end;
+        $chunk_start = $date_end->modify("-{$chunk_size} days");
+        if ($chunk_start < $date_start) {
+            $chunk_start = $date_start;
             $chunk_size = $date_end->diff($date_start)->days;
         }
 
@@ -93,16 +121,16 @@ class Jetpack_Importer {
         $chunk_end = trim($_GET['chunk_end']);
         $chunk_size = (int) trim($_GET['chunk_size']);
         $date_end = new \DateTimeImmutable($params['date-end']);
+        $date_start = new \DateTimeImmutable($params['date-start']);
         $chunk_end = new \DateTimeImmutable($chunk_end);
-        $chunk_start = $chunk_end->modify("-{$chunk_size} days");
 
-        // calculate next chunk end date
-        $next_chunk_end = $chunk_end->modify("+{$chunk_size} days");
-        if ($next_chunk_end > $date_end) {
-            $next_chunk_end = $date_end;
-            $chunk_size = $next_chunk_end->diff($chunk_end)->days;
+        // calculate next chunk end date and actual size of current chunk
+        $next_chunk_end = $chunk_end->modify("-{$chunk_size} days");
+        if ($next_chunk_end < $date_start) {
+            $chunk_size = $chunk_end->diff($date_start)->days;
         }
 
+        // import this chunk
         if (! $this->perform_chunk_import($params['wpcom-api-key'], $params['wpcom-blog-uri'], $chunk_end, $chunk_size)) {
             delete_option('koko_analytics_jetpack_import_params');
             wp_safe_redirect(get_admin_url(null, '/index.php?page=koko-analytics&tab=jetpack_importer&success=0'));
@@ -110,7 +138,7 @@ class Jetpack_Importer {
         }
 
         // If we're done, redirect to success page
-        if ($next_chunk_end == $chunk_end) {
+        if ($next_chunk_end < $date_start) {
             delete_option('koko_analytics_jetpack_import_params');
             wp_safe_redirect(get_admin_url(null, '/index.php?page=koko-analytics&tab=jetpack_importer&success=1'));
             exit;
@@ -118,16 +146,18 @@ class Jetpack_Importer {
 
         $url = add_query_arg(['koko_analytics_action' => 'jetpack_import_chunk', 'chunk_size' => $chunk_size, 'chunk_end' => $next_chunk_end->format('Y-m-d'), '_wpnonce' => wp_create_nonce('koko_analytics_jetpack_import_chunk')]);
 
+        $chunk_start = $chunk_end->modify("-{$chunk_size} days");
+        $chunks_left = (int) ceil($chunk_end->diff($date_start)->days / $chunk_size);
+
         // we could do a wp_safe_redirect() here
         // but instead we send some HTML to the client and perform a client-side redirect just so the user knows we're still alive and working
-        // TODO: Calculate number of steps / progress
-        // TODO: Calculate est. time left? (Assume 1-2seconds per chunk)
         ?>
         <style>body { background: #f0f0f1; color: #3c434a; font-family: sans-serif; font-size: 16px; line-height: 1.5; padding: 32px; }</style>
-        <meta http-equiv="refresh" content="2; url=<?php echo esc_attr($url); ?>">
+        <meta http-equiv="refresh" content="1; url=<?php echo esc_attr($url); ?>">
         <h1>Liberating your data... Please wait.</h1>
         <p>Importing stats between <strong><?php echo $chunk_start->format('Y-m-d'); ?></strong> and <strong><?php echo $chunk_end->format('Y-m-d'); ?></strong>.</p>
         <p>Please do not close this browser tab while the importer is running.</p>
+        <p>Estimated time left: <?php echo round($chunks_left * 1.5); ?> seconds.</p>
         <?php
         exit;
     }
@@ -138,6 +168,7 @@ class Jetpack_Importer {
         $url = "https://stats.wordpress.com/csv.php?api_key={$api_key}&blog_uri={$blog_uri}&end={$date_end->format('Y-m-d')}&table=postviews&format=json&days={$chunk_size}&limit=-1";
         $response = wp_remote_get($url);
         if (!$response || wp_remote_retrieve_response_code($response) >= 400) {
+            error_log("Koko Analytics - JetPack Importer: received error response from WordPress.com API: " . wp_remote_retrieve_body($response));
             return false;
         }
 
@@ -170,14 +201,20 @@ class Jetpack_Importer {
 
                 $query = $wpdb->prepare("INSERT INTO {$wpdb->prefix}koko_analytics_post_stats (date, id, visitors, pageviews) VALUES (%s, %d, %d, %d) ON DUPLICATE KEY UPDATE visitors = visitors + VALUES(visitors), pageviews = pageviews + VALUES(pageviews);", [$item->date, $postviews->post_id, $postviews->views, $postviews->views]);
                 $wpdb->query($query);
+
+                if ($wpdb->last_error !== '') {
+                   error_log("Koko Analytics - JetPack Importer: database error trying to update site_stats: " . $wpdb->last_error);
+                }
+
             }
 
             // update site stats
             $query = $wpdb->prepare("INSERT INTO {$wpdb->prefix}koko_analytics_site_stats (date, visitors, pageviews) VALUES (%s, %d, %d) ON DUPLICATE KEY UPDATE visitors = visitors + VALUES(visitors), pageviews = pageviews + VALUES(pageviews);", [$item->date, $site_views, $site_views]);
             $wpdb->query($query);
+            if ($wpdb->last_error !== '') {
+                error_log("Koko Analytics - JetPack Importer: database error trying to update site_stats: " . $wpdb->last_error);
+            }
         }
-
-        // TODO: log database errors? Or bail entire import?
 
         return true;
     }